requirements.txt
django~=5.2.1
django-richtextfield~=1.6.2
django-ckeditor~=6.7.3
Pillow~=11.2.1